feat: add SaveSessionWithoutContext and LoadSessionBySessionId
